requirements.txt
finnhub-python==2.4.18
pandas-market-calendars==4.1.4
pytz==2023.3
pytrends==4.9.2